ran migration script on example-project (and $app/env -> $app/environment)
