+++ sites/example-project/src/components/ui/QueryViewer.svelte
@@ -1,13 +1,13 @@
 <script>
   import { slide, blur } from 'svelte/transition';
-  import { dev } from '$app/env';
+  import { dev } from '$app/environment';
   import DataTable from './QueryViewerSupport/QueryDataTable.svelte'
   import ChevronToggle from "./ChevronToggle.svelte"
   import Prism from "./QueryViewerSupport/Prismjs.svelte";
   import {showQueries} from '@evidence-dev/components/ui/stores'
   import CompilerToggle from './QueryViewerSupport/CompilerToggle.svelte';
   import { writable } from 'svelte/store';
-  import { browser } from '$app/env';
+  import { browser } from '$app/environment';
 
   export let queryID; 
   export let pageQueries
