+++ sites/example-project/src/pages/+error.svelte
@@ -1,0 +1,5 @@
+<script>
+	import { page } from "$app/stores";
+</script>
+
+<h2>{$page.error.message}</h2>
