+++ sites/example-project/src/pages/api/[route].json/+server.js
@@ -1,12 +1,9 @@
 import {runQueries} from '@evidence-dev/db-orchestrator'
-import { dev } from '$app/env';
+import { dev } from '$app/environment';
+import { json } from '@sveltejs/kit';
 
 export async function GET({params}) {
   const { route } = params;
   const data = await runQueries(route, dev);
-  return {
-      body: {
-        data
-      }
-  };
+  return json({ data })
 }
