+++ sites/example-project/src/pages/api/queryStatus.json/+server.js
@@ -1,23 +1,15 @@
 import fs from 'fs';
-import { dev } from '$app/env';
+import { dev } from '$app/environment';
 
 export async function GET() {
     if (!dev) {
-        return {
-            status: 404
-        }
+        return new Response(undefined, { status: 404 })
     }
     else {
         let queryProgress = {}
         if (fs.existsSync('.evidence-queries/status.json')) {
             queryProgress = JSON.parse(fs.readFileSync('.evidence-queries/status.json', 'utf8'));
         }
-        return {
-            header: "accept: application/json",
-            status: 200,
-            body: {
-                queryProgress
-            }
-        }
+        return json({ queryProgress })
     }
 }
