+++ sites/example-project/src/pages/api/settings.json/+server.js
@@ -1,8 +1,9 @@
 import fs from 'fs';
 import path from 'path';
 import gitRemoteOriginUrl from 'git-remote-origin-url'; // get the git repo
-import { dev } from '$app/env';
+import { dev } from '$app/environment';
 import logEvent from '@evidence-dev/telemetry'; 
+import { json } from '@sveltejs/kit';
 
 function getLocalGitRepo () {
     if(fs.existsSync(path.join(path.resolve('../../'), '.git'))){
@@ -12,9 +13,7 @@ function getLocalGitRepo () {
 
 export async function GET() {
     if (!dev) {
-        return {
-            status: 404
-        }
+        return new Response(undefined, { status: 404 })
     }
     else { 
         let settings = {}
@@ -31,14 +30,7 @@ export async function GET() {
         }catch {
             
         }
-        return {
-            header: "accept: application/json",
-            status: 200,
-            body: {
-                settings,
-                gitIgnore
-            }
-        }
+        return json({ settings, gitIgnore })
     }
 }
 
@@ -83,9 +75,7 @@ export async function POST({request}) {
             fs.writeFileSync('../../.gitignore', gitIgnore)
         }
     }
-    return {
-        body: settings
-    }
+    return json(settings)
 }
 
 // Breaking changes in new verion of Svelte kit - merged on Jan 19, 2022
