+++ sites/example-project/src/pages/charts/area-chart/+page.md
