+++ sites/example-project/src/pages/charts/bar-chart/+page.md
