+++ sites/example-project/src/pages/charts/histogram/+page.md
