+++ sites/example-project/src/pages/charts/mixed-charts/+page.md
