+++ sites/example-project/src/pages/charts/stacked-charts/+page.md
