+++ sites/example-project/src/pages/settings/+page.svelte
@@ -1,33 +1,9 @@
-<script context="module">
-  import { dev } from "$app/env";
-  export const load = async ({ fetch }) => {
-    if (dev) {
-      const settingsRes = await fetch("../api/settings.json");
-      const customFormattingSettingsRes = await fetch("../api/customFormattingSettings.json");
-      const { settings, gitIgnore } = await settingsRes.json();
-      const { customFormattingSettings } = await customFormattingSettingsRes.json();
-      return {
-        props: {
-          settings,
-          gitIgnore,
-          customFormattingSettings,
-        },
-      };
-    } else {
-      return {
-        props: {
-          settings: {},
-          gitIgnore: "",
-        },
-      };
-    }
-  };
-</script>
-
 <script>
-  export let settings;
-  export let customFormattingSettings;
-  export let gitIgnore;
+  export let data;
+  let { settings, customFormattingSettings, gitIgnore } = data;
+  $: ({ settings, customFormattingSettings, gitIgnore } = data);
+
+  import { dev } from "$app/environment";
   import DatabaseSettingsPanel from "@evidence-dev/components/ui/Databases/DatabaseSettingsPanel.svelte";
   import VersionControlPanel from "@evidence-dev/components/ui/VersionControl/VersionControlPanel.svelte";
   import DeploySettingsPanel from "@evidence-dev/components/ui/Deployment/DeploySettingsPanel.svelte";
