+++ sites/example-project/src/pages/status/+page.svelte
