+++ sites/example-project/src/pages/text-and-metrics/value-component/+page.md
